requirements.txt
prettytable==3.12.0
PyMySQL==1.1.1
python-dotenv==1.0.1
wcwidth==0.2.13
